sdl image is such a mess compared to plain sdl -.-
is it worth it?

my initial plan was to go full SDL everywhere and everything
sdl for sdl
sdl for image loading
sdl for networking
sdl for threading
sdl for typedefs
sdl for logging

using SDL, SDL_Net and SDL_Image for now and potentially more

But maybe only SDL is the super well maintained highly used library
that is actually fun as a dependency.
And the other sdl projects are at least as much of a pain to depend
on as something outside of the sdl universe.